# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Simple378 Fraud Detection API
  description: |
    Enterprise-grade fraud detection and investigation platform API.

    ## Authentication
    All API endpoints require authentication via JWT tokens obtained from `/auth/login`.

    ## Rate Limiting
    - 100 requests per minute per IP
    - 1000 requests per minute for authenticated users
    - Progressive backoff on violations

    ## Security
    - All data encrypted at rest using SQLCipher
    - HMAC-SHA256 signed IPC communication
    - Comprehensive audit logging
    - Zero-trust architecture

    ## Versioning
    API versioning follows semantic versioning with breaking changes in major versions.
  version: 1.0.0
  contact:
    name: Simple378 Support
    email: support@378x492.com
  license:
    name: Proprietary
    url: https://378x492.com/license
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.378x492.com/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /cases:
    get:
      summary: List cases with filtering and pagination
      description: Retrieve paginated list of fraud cases with advanced filtering
      tags: [Cases]
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, open, investigating, pending_review, escalated, closed_approved, closed_denied, closed_no_action]
          description: Filter by case status
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
          description: Filter by case priority
        - name: risk_score_min
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 100
          description: Minimum risk score filter
        - name: assigned_to
          in: query
          schema:
            type: string
          description: Filter by assigned investigator
        - name: search
          in: query
          schema:
            type: string
          description: Full-text search across case data
      responses:
        '200':
          description: Cases retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Case'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  filters:
                    type: object
                    description: Applied filters summary
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create new case
      description: Create a new fraud investigation case
      tags: [Cases]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - priority
              properties:
                title:
                  type: string
                  maxLength: 200
                  description: Case title
                  example: Suspicious wire transfer pattern
                description:
                  type: string
                  description: Detailed case description
                  example: Multiple wire transfers just below $10k threshold
                priority:
                  type: string
                  enum: [low, medium, high, critical]
                  default: medium
                type:
                  type: string
                  enum: [fraud_suspected, identity_theft, account_takeover, money_laundering]
                  default: fraud_suspected
                assigned_to:
                  type: string
                  description: User ID to assign case to
                tags:
                  type: array
                  items:
                    type: string
                  description: Case tags for categorization
      responses:
        '201':
          description: Case created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Case'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cases/{caseId}:
    get:
      summary: Get case details
      description: Retrieve detailed information about a specific case
      tags: [Cases]
      security:
        - BearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique case identifier
      responses:
        '200':
          description: Case details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/CaseDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update case
      description: Update case information and status
      tags: [Cases]
      security:
        - BearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 200
                description:
                  type: string
                status:
                  type: string
                  enum: [draft, open, investigating, pending_review, escalated, closed_approved, closed_denied, closed_no_action]
                priority:
                  type: string
                  enum: [low, medium, high, critical]
                assigned_to:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Case updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Case'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete case
      description: Soft delete a case (mark as archived)
      tags: [Cases]
      security:
        - BearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Case deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Case archived successfully

  # Evidence Management
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
        role:
          type: string
          enum: [admin, investigator, analyst, auditor]
        permissions:
          type: array
          items:
            type: string
        lastLogin:
          type: string
          format: date-time
        isActive:
          type: boolean
...
    Case:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [draft, open, investigating, pending_review, escalated, closed_approved, closed_denied, closed_no_action]
        priority:
          type: string
          enum: [low, medium, high, critical]
        riskScore:
          type: number
          minimum: 0
          maximum: 100
        type:
          type: string
          enum: [fraud_suspected, identity_theft, account_takeover, money_laundering]
        assignedTo:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
    CaseDetail:
      allOf:
        - $ref: '#/components/schemas/Case'
        - type: object
          properties:
            description:
              type: string
            evidence:
              type: array
              items:
                $ref: '#/components/schemas/Evidence'
            timeline:
              type: array
              items:
                $ref: '#/components/schemas/TimelineEvent'
            relatedCases:
              type: array
              items:
                type: string
                format: uuid
            aiAnalysis:
              type: object
              description: AI-generated insights and recommendations
    Evidence:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        hash:
          type: string
          description: SHA-256 file hash
        uploadedAt:
          type: string
          format: date-time
        uploadedBy:
          $ref: '#/components/schemas/User'
        caseId:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
        analysis:
          type: object
          properties:
            ocr:
              type: object
              description: OCR extraction results
            metadata:
              type: object
              description: File metadata
            fraudIndicators:
              type: array
              items:
                type: string
              description: Detected fraud indicators
    TimelineEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        type:
          type: string
          enum: [case_created, evidence_uploaded, analysis_completed, status_changed, comment_added, alert_triggered]
        description:
          type: string
        user:
          $ref: '#/components/schemas/User'
        metadata:
          type: object
          description: Event-specific data
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...
  responses:
    BadRequest:
      description: Bad request - invalid input parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Invalid input parameters
              details:
                type: object
                description: Validation error details
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Authentication required
              code:
                type: string
                example: UNAUTHORIZED
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Resource not found
              code:
                type: string
                example: NOT_FOUND
...